pyproject.toml
[project]
name = "h-2a-paper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "addfips>=0.4.2",
    "altair>=6.0.0",
    "bs4>=0.0.2",
    "chardet>=5.2.0",
    "docling>=2.76.0",
    "dspy>=3.1.3",
    "exactextract>=0.3.0",
    "fastexcel>=0.19.0",
    "fastparquet>=2024.11.0",
    "formulaic>=1.2.1",
    "geopandas>=1.1.2",
    "google-genai>=1.56.0",
    "google-maps-places>=0.5.0",
    "googlemaps>=4.10.0",
    "great-tables>=0.20.0",
    "ibis-framework[duckdb,polars]>=12.0.0",
    "ipykernel>=6.29.5",
    "joblib>=1.5.3",
    "jupyter>=1.1.1",
    "lxml>=6.0.2",
    "marimo[recommended]>=0.18.4",
    "matplotlib>=3.10.8",
    "nasspython>=1.0.0",
    "nbconvert>=7.16.6",
    "notebook>=7.4.4",
    "numba>=0.64.0",
    "numpy>=2.3.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pdfplumber>=0.11.7",
    "pip>=25.1.1",
    "polars>=1.36.1",
    "polars-io>=0.5.0",
    "pyarrow>=20.0.0",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.47.0",
    "pyfixest>=0.40.0",
    "pyprojroot>=0.3.0",
    "python-dotenv>=1.2.1",
    "python-lsp-ruff>=2.3.0",
    "python-lsp-server>=1.14.0",
    "rapidfuzz>=3.13.0",
    "rasterio>=1.5.0",
    "ratelimit>=2.2.1",
    "requests>=2.32.4",
    "rioxarray>=0.21.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "tabulate>=0.9.0",
    "tqdm>=4.67.3",
    "us>=3.2.0",
    "vegafusion>=2.0.3",
    "vl-convert-python>=1.9.0.post1",
    "websockets>=15.0.1",
    "xlrd>=2.0.2",
    "xlsx2csv>=0.8.6",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "pymupdf>=1.27.1",
    "jax[cuda13]>=0.9.1",
    "optax>=0.2.7",
    "lineax>=0.1.0",
    "scikit-learn>=1.8.0",
    "jaxlib[cuda12,cuda13]>=0.9.2",
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
